refactor(reuse): dedup fmtRelative em @/Lib/datetime-br (gate reuse verde)
fmtRelative estava duplicado e exportado em dois lugares, criando uma
duplicata NOVA no ratchet reuse-index (util:fmtRelative 2×) que deixava
o gate `reuse:duplicates` vermelho em PRs subsequentes:

  - resources/js/Pages/kb/_lib/helpers.ts:62
  - resources/js/Pages/team-mcp/CcSessions/_components/sessionTokens.ts:30
    (entrou no main via #2821 / commit 14d094929, sem reusar nem baselinar)

Extrai a versao canonica (robusta, com guard de NaN pra seeds relativos)
pra resources/js/Lib/datetime-br.ts e re-exporto dos dois modulos. Os
consumidores (kb HealthPanel/NodeReader, CcSessions/Index) ficam intactos
e o re-export nao e indexado pelo reuse-index -> gate verde sem mexer no
baseline.

Validado: node scripts/reuse-index.mjs --gate (nenhuma duplicata nova).

Refs: reuse-gate ratchet . MANUAL-CSS-JS #5 . ADR 0240
